fix(build): keep HTTP admission lazy
